chore: Separate sources from site for branch-based publishing.
This commit makes changes to a new workflow, designed to declutter
commit logs by separating out the generated site files into their own
branch. The result is that the built site can no longer (easily) be
committed as a part of changes intended to only modify the source files.

Specifically, this commit:
  - Moves source files from ./src to .
  - Removes ./docs/* from the main branch
  - Decouples the locally-served build from the publishable build by
    separating out their build directories